Updated project: moved view files to new directory
