+++ voca-frontend/src/App.js
@@ -3,9 +3,9 @@ import "./App.css";
 import { useEffect, useState } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import React from "react";
-import MainPage from "./MainPage";
-import Signup from "./Signup";
-import Signin from "./Signin";
+import MainPage from "./component/MainPage";
+import Signup from "./component/Signup";
+import Signin from "./component/Signin";
 
 function App() {
   const [data, setData] = useState([]);
